feat: Add use-top-artists hook for fetching top artists
This commit adds a new hook called use-top-artists, which is responsible for fetching the top artists from the highest context. It uses the useSWR hook from the swr library to handle the data fetching. The fetched data is then returned as part of the hook's return value, along with isLoading and error states.

The use-top-artists hook is used in the GuessInput component to populate the options for the Autocomplete component. This allows users to select an artist from the fetched data.

This commit also includes the addition of the Artist and ArtistsResponse interfaces in the types/artists.ts file, which define the shape of the artist data returned from the API.